+++ publish/CHANGELOG.md
@@ -2,6 +2,7 @@
   - Adds a new parameter "day" to get the current day of the world.
   - Adds a new parameter "ticks" to get ticks since the world start.
   - Adds a new parameter "pos" to get the object position as x,z,y.
+  - Adds new data fields `position` and `rotation`.
   - Changes the parameter "<time>" to be seconds instead of ticks.
   - Adds a new field `packaged` to RPC calls to send parameters as a single package (some RPCs require this).
   - Adds support for the data field `connection` to use ZDO id instead of connection hash.